+++ tests/Db/QueryBuilder/Condition/Builder/LikeBuilderTest.php
@@ -4,11 +4,18 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Tests\Db\QueryBuilder\Condition\Builder;
 
+use PHPUnit\Framework\Attributes\TestWith;
 use PHPUnit\Framework\TestCase;
 use InvalidArgumentException;
+use Yiisoft\Db\Command\Param;
+use Yiisoft\Db\Constant\DataType;
 use Yiisoft\Db\QueryBuilder\Condition\Builder\LikeBuilder;
 use Yiisoft\Db\QueryBuilder\Condition\Like;
+use Yiisoft\Db\QueryBuilder\Condition\LikeMode;
 use Yiisoft\Db\Tests\Support\TestTrait;
+
+use function PHPUnit\Framework\assertCount;
+use function PHPUnit\Framework\assertSame;
 
 /**
  * @group db
@@ -27,5 +34,24 @@ final class LikeBuilderTest extends TestCase
         $this->expectExceptionMessage('Invalid operator in like condition: "INVALID"');
 
         (new LikeBuilder($db->getQueryBuilder()))->build($likeCondition);
+    }
+
+    #[TestWith(['%test%', LikeMode::Contains])]
+    #[TestWith(['test%', LikeMode::StartsWith])]
+    #[TestWith(['%test', LikeMode::EndsWith])]
+    public function testBuildWithContainsMode(string $expected, LikeMode $mode): void
+    {
+        $likeCondition = new Like('column', 'LIKE', 'test', mode: $mode);
+        $likeBuilder = new LikeBuilder($this->getConnection()->getQueryBuilder());
+
+        $params = [];
+        $likeBuilder->build($likeCondition, $params);
+
+        assertCount(1, $params);
+
+        /** @var Param $param */
+        $param = reset($params);
+        assertSame($expected, $param->value);
+        assertSame(DataType::STRING, $param->type);
     }
 }